requirements.txt
agno==1.4.2
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.16
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.18.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.39.0
google-genai==1.12.1
gradio==5.27.0
gradio_client==1.9.0
groovy==0.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.30.2
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
lancedb==0.22.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
numpy==2.2.5
openai==1.76.0
orjson==3.10.16
overrides==7.7.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydub==0.25.1
Pygments==2.19.1
pypdf==5.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
rich==14.0.0
rsa==4.9.1
ruff==0.11.7
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
tantivy==0.22.2
tomli==2.2.1
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
